libs: nu32: introduces nu32_init_leds() and nu32_init_serial() functions
There are now three global objects and corresponding pointers in nu32.* to
facilitate easier use of the leds and serial port on the nu32. Note that
nu32_led0p and nu32_led1p start out NULL so that if nu32_init_leds() fails,
it can be determined which led failed to init.