Mostly directory reorganization
This is to promote code reuse. All include directives for NUSolar API
headers should have the filename prefixed with "nu/" for the C API
or with "nupp/" for the C++ API. This allows unambiguous usage of the
C API files inside the C++ API.

Additionally, there are some minor edits and various edits to the C API of which unfortunately
I can't recall all the details.