move pbus_hz() to param.hpp interface
it's used in spi and such which can presume a peripheral bus clock. in
cases where the clock's the same, can't we just define pbus_hz() to
Hz()? I added a generic param.h which illustrates this

Also we can configure the Pbus clock right? or is it a function of the
main clock's frequency?