Compile-time checking of timer arguments
Conversion and delay functions/macros are now compile-time checked
if their arguments are constant expressions. Delays are limited to
53s for nu_delay_s() and 5300 for all other delay functions
(nu_delay_ms, nu_delay_us, nu_delay_ns). Same for conversions.
Keep in mind that expressions that are NOT compile-time constants
will not be checked at compile-time OR runtime, so don't do that
please.

Also put everything in the nu namespace and some other minor edits.