enable stack allocation for stringbuf
each projects allocates 4096 bytes on the stack for use of SPI and
serial streaming. hopefully we never stream more than 4kb at once.

Currently deallocation is a problem. It only succeeds if things are
deallocated in precisely the reverse order of allocation.

running git clear is sometimes needed to add source files to mplab.