partially resolve symbol loading issues
steering_wheel and driver_controls now load properly in debug. BMS
doesn't, but it's because of ad7685 (don't know why). commenting this
out fixes the bms.

unfortunately… this is the return of MAXIMUM WARNING MAXIMUM ERROR.
stream crap has to be included very early (varies based on the project)

the release builds still don't load symbols. it's directly related to
library optimization, as setting Optimization=0 fixes it. It's probably
one of the custom options